(fine) Update sitemap generator to deal with new `url` semantics for page objects
At some point, they harmonized the contents of the `url` variable in page and
post objects, so the sitemap generator doesn't need to append `@dir` anymore.
Without this patch, the sitemap URLs contained duplicated directories for each
page.

As of this commit, the site generates just fine, assuming our branch includes
the fix for the force_encoding Kramdownpocalypse.