Switch to Redcarpet for Markdown processing
HOLY CROW, that is more than twice as fast for a complete site build. We're
sub-one-minute for the first time in years.

We should also be able to use fenced code blocks with language specifiers now;
that's to be tested in a later commit.

Blemishes: The smartypants implementation doesn't smarten up apostrophes in
contractions, for some reason. That's about it.

We shouldn't need to add Redcarpet to the Gemfile, as Jekyll has adopted it as
the default for new sites and will automatically install an appropriate version
as a dependency. (i.e. it won't try to install redcarpet 3.)