Re-enable footnotes with unreleased Redcarpet version (advertises self as 2.3.0)
Wow, we got lucky here -- the footnote commits get us a bunch of improvements
over 2.2.2, without breaking anything, and the code still advertises itself as
being version 2.3.0. In summary:

- Footnotes are a go.
- The bug with SmartyPants leaving a bunch of unsmartened apostrophes and single
  quotes is fixed.
- A bug with not processing emphasis inside parentheses is fixed.
- Alas, it also surfaced some cosmetic problems in the MCollective docs, where
  intra-word code spans were being used to suppress intra-emphasis. I fixed
  these in marionette-collective commit b18d9994e687b7a3520e90d5a283eb775ff0e865.

This commit adds Redcarpet as a git gem, pinned to commit 40923d2ba -- note we
put it in front of Jekyll in the Gemfile, so we don't accidentally install two
versions. It also adds "footnotes" back to the Redcarpet extensions in
_config.yml.

This means we're carrying a patch on Jekyll again, though it's only a change to
the gemspec. Specifically, we depend on this diff:

diff --git a/jekyll.gemspec b/jekyll.gemspec
index b29d966..a376bfe 100644
--- a/jekyll.gemspec
+++ b/jekyll.gemspec
@@ -32,7 +32,7 @@ Gem::Specification.new do |s|
   s.add_runtime_dependency('commander', "~> 4.1.3")
   s.add_runtime_dependency('safe_yaml', "~> 0.7.0")
   s.add_runtime_dependency('colorator', "~> 0.1")
-  s.add_runtime_dependency('redcarpet', "~> 2.2.2")
+  s.add_runtime_dependency('redcarpet', "~> 2.2")

   s.add_development_dependency('rake', "~> 10.0.3")
   s.add_development_dependency('rdoc', "~> 3.11")

This is in the aug_2013_upgrade_test branch as commit a1f56495.

NOTE: The only remaining problem with switching at this point is that Redcarpet
does not support the definition list extension, which we use extensively through
the docs. This may be a blocker for now, until we can modify the formatting of
the type reference, among other things. Kids, learn your lesson from us: be
parsimonious about Markdown extensions, because they'll always bite you in the
ass eventually.