Update Jekyll to 2.0.x and Redcarpet to Whatever; disable the recent_commits tag
Hey, looks like the stock versions work more or less fine now! (Now that I'm on
a Mac OS that has Ruby 2.)

As a complication, it looks like the recent_commits tag is blowing up for some
reason and preventing the site from generating. It doesn't seem to provide much
value anyway, so let's write it off as a failed experiment and move on.